# src/main/resources/swagger/example/openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "description": "Development server",
      "url": "http://localhost:8080/api"
    }
  ],
  "info": {
    "description": "This is a simple example API for Helpdesk",
    "version": "1.0.0",
    "title": "Simple Helpdesk API",
    "contact": {
      "email": "mentor@novahub.vn"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/admin/users/{id}": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Update user by id (ADMIN)",
        "description": "Update user by id (ADMIN)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          },
          {
            "in": "header",
            "name": "access_token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "dff4d0d4891bfd8bafdb859dac4cbccb218bcf5adc9887a627c6c159f76ef0b9"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string",
                    "description": "new password (optional)",
                    "example": "password"
                  },
                  "first_name": {
                    "type": "string",
                    "example": "help"
                  },
                  "last_name": {
                    "type": "string",
                    "example": "desk"
                  },
                  "birth_day": {
                    "type": "string",
                    "description": "yyyy-MM-dd",
                    "example": "1997-09-01"
                  },
                  "address": {
                    "type": "string",
                    "example": "14 An Thuong 32 street, Da Nang city"
                  },
                  "phone": {
                    "type": "string",
                    "example": "0905123456"
                  },
                  "title": {
                    "type": "string",
                    "example": "Back-end Developer"
                  },
                  "introduction": {
                    "type": "string",
                    "example": "I work as a back-end developer"
                  },
                  "avatar_url": {
                    "type": "string",
                    "example": "http://localhost:8080/avatar.png"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update user by id successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account",
                  "properties": {
                    "role": {
                      "$ref": "#/components/schemas/Role"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Token is expired",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": 1532589762.947
                    },
                    "status": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Token is expired"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "TokenIsExpiredException with token = 579989c508208d49c978e6cb99383688e6f1fedabfc22c27d3542ab0341d3b13"
                    },
                    "path": {
                      "type": "string",
                      "example": "/admin/users/{id}"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532506012.933"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 401
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Unauthorized"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid token"
                    },
                    "path": {
                      "type": "string",
                      "example": "/admin/users/{id}"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access is denied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532506012.933"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 403
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Access is denied"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Access is denied"
                    },
                    "path": {
                      "type": "string",
                      "example": "/admin/users/{id}"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532462094.006"
                    },
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "User not found"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "UserNotFoundException with id = 1"
                    },
                    "path": {
                      "type": "string",
                      "example": "/admin/users/{id}"
                    }
                  }
                }
              }
            }
          },
          "406": {
            "description": "Not acceptable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532506012.933"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 406
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "password": {
                          "type": "string",
                          "enum": [
                            "password must have between 8 and 40 characters"
                          ],
                          "example": "password must have between 8 and 40 characters"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "AccountValidationException with number of errors = 1"
                    },
                    "path": {
                      "type": "string",
                      "example": "/admin/users/{id}"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete user by id (ADMIN)",
        "description": "Delete user by id (ADMIN)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          },
          {
            "in": "header",
            "name": "access_token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "dff4d0d4891bfd8bafdb859dac4cbccb218bcf5adc9887a627c6c159f76ef0b9"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete user by id successfully"
          },
          "400": {
            "description": "Token is expired",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": 1532589762.947
                    },
                    "status": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Token is expired"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "TokenIsExpiredException with token = 579989c508208d49c978e6cb99383688e6f1fedabfc22c27d3542ab0341d3b13"
                    },
                    "path": {
                      "type": "string",
                      "example": "/admin/users/{id}"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532506012.933"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 401
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Unauthorized"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid token"
                    },
                    "path": {
                      "type": "string",
                      "example": "/admin/users/{id}"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access is denied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532506012.933"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 403
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Access is denied"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Access is denied"
                    },
                    "path": {
                      "type": "string",
                      "example": "/admin/users/{id}"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532462094.006"
                    },
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "User not found"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "UserNotFoundException with id = 1"
                    },
                    "path": {
                      "type": "string",
                      "example": "/admin/users/{id}"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Account": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "email": {
            "type": "string",
            "example": "helpdesk@novahub.vn"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "LOCKED"
            ],
            "example": "ACTIVE"
          },
          "first_name": {
            "type": "string",
            "example": "help"
          },
          "last_name": {
            "type": "string",
            "example": "desk"
          },
          "birth_day": {
            "type": "string",
            "format": "date",
            "description": "pattern yyyy-MM-dd",
            "example": "1997-09-01T00:00:00.000Z"
          },
          "address": {
            "type": "string",
            "example": "14 An Thuong 32 street, Da Nang city"
          },
          "phone": {
            "type": "string",
            "example": 20725
          },
          "title": {
            "type": "string",
            "example": "Junior Front-end"
          },
          "introduction": {
            "type": "string",
            "example": "My name is Than Quang Minh. I come from Vietnam. Currently I work as a web developer. I have more than 15 years programming and more than 5 years of work experience in Software & web development using .NET, PHP, RoR, Java Spring Framework and adequate tutoring & on-job training experience, as well as strong background in Computer Science, computer system, database modeling & information architecture, and fluent in algorithmic & OOP programming and web-related technologies."
          },
          "avatar_url": {
            "type": "string",
            "example": "http://localhost:8080/avatar.png"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2018-07-24 17:50:45"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2018-07-24 17:50:45"
          },
          "role_id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "enum": [
              "ADMIN",
              "CLERK",
              "EMPLOYEE"
            ],
            "example": "ADMIN"
          }
        }
      },
...
    }
  }
}